Fix ignore while creating .gitignore file
Problem:

We're calling git diff on .gitignore before we had the .gitignore file.

Error message:

fatal: ambiguous argument '.gitignore': unknown revision or path not in
the working tree.

Solution:

Now the first step is checking if the .gitignore file exists. If it does
not exists we create and commit it.