Add shfmt-diff to lint-shell make target
When we just use the shfmt-edit target we'll edit the files to comply
with the rules of shfmt but we may not notice that we had changes to
files.

To make it easier to identify file changes, we added shfmt-diff to
lint-shell. With this change the shfmt-diff will print the diff to make
it clear that we have changes to some files due to shfmt.